fix: svg export only works in the XY plane
The export() method in Mod/Draft/importSVG.py assumes it is looking down from Z onto the XY plane.
This works fine to export shapes parallel to that plan.

For flat or mostly flat objects that are aligned to the XZ or YZ plane, this exports only a thin line, as it looks sideways on the object.
A general solution to this problem would e.g. use the draft WorkingPlane (or the current viewing direction).

This PR implements a minimalistic solution that works very well for exporting sketches oriented in one of the main three planes.
We simply check which edge of the bounding box is shortest and construct a temporary working plane normal to that direction.

My usecase is 2D Designs for plotting or lasercutting. For this, the simple solution is just perfect, as I do not need to bother about the plan that a certain sketch was started at all.